fix: preserve validation across file moves
